feat(saf): Implement minimal SAF reader
Support for fairly bare-bones SAF reading, including individual readers
for the index file (".saf.idx"), the position file (".pos.saf.gz"), and
the values file (".saf.gz"), as well as a joint reader struct to
incorporate them. Reading can be done with iterators over the value and
positions files individually or jointly, and reading values to a 2d
ndarray is also supported when the optional dependency on ndarray is
enabled, as it is by default.